WIP: Honor lock file requirements for packages `composer`, `composer-plugin-api` and `composer-runtime-api` during platform package resolution
Packages (whether it's the root `composer.json` or a dependency) can depend on explicit versions of virtual packages named 'composer', 'composer-plugin-api', or 'composer-runtime-api'. This is important, for example, for Composer plugins (https://getcomposer.org/doc/articles/plugins.md), which may only be compatible with certain plugin API versions, or for libraries that use the Composer runtime API (https://getcomposer.org/doc/07-runtime.md).

Any such requirement needs to be included by us for overall solving, since it may impact what Composer version gets installed - right now, in practice, this mostly affects 2.2-LTS plugins (e.g. Magento), and we are special casing <=2.2 in our plugin API selection logic (which uses the lock file's 'plugin-api-version' field), so for the currently-common cases, it works fine, but there will be API changes again in the future that plugins or packages may not be compatible with, and we need the ability to more precisely control what version or series of Composer can get installed.

TODO: more tests, e.g. for picking an older version explicitly, in the composer fixtures dir.