WIP: human-readable output to alternate FD inside installer plugin
Instead of grepping and stripping on the caller side, we are now passing an additional file descriptor to the plugin, and it writes the "short", human-readable output (just package names and versions with, for now, "  - " prefixed) of package installs and extension activations.

The regular full output from Composer and the plugin still goes to install.log.