Stronger isolation for bootstrapped Composer and PHP
With this change, the bootstrapped PHP version is used until the very end even for the APM auto-installs. Avoids the need for trickery around PHP_INI_SCAN_DIR to prevent already-installed APM extension startups etc, and guarantees we're running a version of PHP the installer plugin is compatible with.

Moves it a little closer to how things work in CNB land, too ;)

GUS-W-18988200